requirements.txt
GitPython==0.3.2.RC1
async==0.6.1
gitdb==0.5.4
github3.py==0.9.0
gnureadline==6.3.3
ipdb==0.8
ipython==2.1.0
pygithub3==0.5
requests==2.3.0
smmap==0.8.2
uritemplate.py==0.3.0
wsgiref==0.1.2
git+https://github.com/ministryofjustice/cotton.git
